feat: 엔티티에 홈, 유저 연관관계 추가 #8
